+++ Assets/Script/GamePlay/Skills/TimeSkillBase.cs
@@ -25,7 +25,7 @@ namespace Assets.Script.GamePlay.Skills
 
 		}
 
-		public bool ActivateSkill(float power,Hero hero,SHeroAnimator heroAnimator)
+		public bool ActivateSkill(int power,Hero hero,SHeroAnimator heroAnimator)
 		{
 			MyHero = hero;
 			HeroAnimator = heroAnimator;
